refactor: Migrate core types directly into pane_state module
- Moved EditorMode, Pane, and PaneCapabilities from core/types.rs to pane_state/mod.rs
- Removed unnecessary core subdirectory
- Updated all imports to reference types from pane_state module directly
- Simplified module hierarchy by eliminating extra nesting

This consolidates pane-related types in their logical location and reduces
unnecessary directory depth. All tests passing (476 unit tests).
