--- src/repl/models/pane_state/core/mod.rs
@@ -1,5 +1,0 @@
-//! Core types for pane state
-
-mod types;
-
-pub use types::{EditorMode, LogicalPosition, LogicalRange, Pane, PaneCapabilities};
