+++ src/repl/models/pane_state/mod.rs
@@ -21,16 +21,18 @@
 //! Previously scattered across multiple classes, this consolidation improves maintainability
 //! and follows the Single Responsibility Principle.
 
-use crate::repl::events::{EditorMode, LogicalPosition, Pane, PaneCapabilities};
 use crate::repl::models::coordinates::geometry::{Dimensions, Position};
 use crate::repl::models::Selection;
 use crate::repl::models::{BufferModel, DisplayCache};
+use bitflags::bitflags;
 use std::ops::{Index, IndexMut};
+
+// Re-export logical position types from models for backward compatibility
+pub use crate::repl::models::{LogicalPosition, LogicalRange};
 
 // Re-export all modules
 pub mod capabilities;
 pub mod content;
-pub mod core;
 pub mod cursor_basic;
 pub mod cursor_line;
 pub mod display;
@@ -43,6 +45,97 @@ pub mod word_navigation;
 
 // Re-export key types for external use
 pub use visual_selection::VisualSelectionRestoreResult;
+
+/// Which pane is currently active
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum Pane {
+    Request,
+    Response,
+}
+
+/// Editor mode (vim-style)
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum EditorMode {
+    Normal,
+    Insert,
+    Command,
+    /// G prefix mode - waiting for second character after 'g' press
+    GPrefix,
+    /// D prefix mode - waiting for second character after 'd' press
+    DPrefix,
+    /// Y prefix mode - waiting for second character after 'y' press
+    YPrefix,
+    /// Visual mode - character-wise text selection mode (vim's 'v')
+    Visual,
+    /// Visual Line mode - line-wise text selection mode (vim's 'V')
+    VisualLine,
+    /// Visual Block mode - block-wise text selection mode (vim's Ctrl+V)
+    VisualBlock,
+    /// Visual Block Insert mode - special insert mode for Visual Block 'I' and 'A' commands
+    VisualBlockInsert,
+}
+
+bitflags! {
+    /// Capabilities that control what operations are allowed on a pane
+    ///
+    /// This bitflag enum provides fine-grained control over pane functionality,
+    /// allowing for flexible configuration without hardcoding pane-specific behavior.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use blueline::repl::events::PaneCapabilities;
+    ///
+    /// // Request pane with full access
+    /// let request_caps = PaneCapabilities::FULL_ACCESS;
+    ///
+    /// // Response pane (read-only)
+    /// let response_caps = PaneCapabilities::READ_ONLY;
+    ///
+    /// // Custom configuration
+    /// let custom_caps = PaneCapabilities::FOCUSABLE | PaneCapabilities::NAVIGABLE;
+    ///
+    /// // Check capabilities
+    /// if request_caps.contains(PaneCapabilities::EDITABLE) {
+    ///     // Allow editing operations
+    /// }
+    /// ```
+    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+    pub struct PaneCapabilities: u32 {
+        /// No capabilities - pane is completely inactive
+        const NONE         = 0b00000000;
+
+        /// Can receive focus and become the active pane
+        const FOCUSABLE    = 0b00000001;
+
+        /// Can edit content (insert, delete, modify text)
+        const EDITABLE     = 0b00000010;
+
+        /// Can select text for visual operations
+        const SELECTABLE   = 0b00000100;
+
+        /// Can scroll content vertically and horizontally
+        const SCROLLABLE   = 0b00001000;
+
+        /// Can navigate with cursor movement commands
+        const NAVIGABLE    = 0b00010000;
+
+        /// Standard read-only configuration for display panes
+        /// Allows focus, navigation, selection, and scrolling but not editing
+        const READ_ONLY = Self::FOCUSABLE.bits()
+                        | Self::SCROLLABLE.bits()
+                        | Self::NAVIGABLE.bits()
+                        | Self::SELECTABLE.bits();
+
+        /// Full access configuration for editable panes
+        /// Enables all capabilities for complete pane functionality
+        const FULL_ACCESS = Self::FOCUSABLE.bits()
+                          | Self::EDITABLE.bits()
+                          | Self::SELECTABLE.bits()
+                          | Self::SCROLLABLE.bits()
+                          | Self::NAVIGABLE.bits();
+    }
+}
 
 /// Minimum width for line number column as specified in requirements
 const MIN_LINE_NUMBER_WIDTH: usize = 3;
